Un-fix Python location
Prior to patch 7.3.755, Vim's configure script looked for Python 2 under
the name `python`. This didn't work on Arch, because Arch's
`/usr/bin/python` is Python 3. Thus, the PKGBUILD patched the configure
script to look for `python2` instead.

After patch 7.3.755, the upstream configure script is smarter; it looks
for `python2` before `python`. So now, Arch's patch actually breaks the
configure script.

This removes the editing of the configure script from the PKGBUILD, now
relying on the upstream configure to locate Python correctly.